refactor of cluster utility functions to support backrest using them
